0503Finish threshold_l setting

+++ client/client_shared.h
@@ -115,7 +115,7 @@ struct mosq_config {
 	int sub_opts; /* sub */
 	long session_expiry_interval;
 	int random_filter; /* sub */
-	int threshold_l;	/*sub*/ //20230412 threshold_l for every client (subscriber)
+	uint16_t threshold_l;	/*sub*/ //20230412 threshold_l for every client (subscriber)
 #ifdef WITH_SOCKS
 	char *socks5_host;
 	int socks5_port;
